Do not break when errors occur; remove gluster pro
For some reason this code had a break statement on line 159 that
does not appear to be correct, since it would end that greenlet
from participating in the rest of the requests.

The hook to invoke the gluster profile command was removed, since
it was hard-coded and is not a general swift tool.

Signed-off-by: Peter Portante <peter.portante@redhat.com>